# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: AgroTech System API
  description: API para monitoramento inteligente em agricultura de precisão. Inclui gestão de usuários, áreas, sensores, leituras, regras e alertas.
  version: 1.0.0
  contact:
    name: AgroTech Team
servers:
  - url: http://localhost:8080/api
    description: Servidor de desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Rule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        operator:
          type: string
          enum: [">", "<", ">=", "<=", "=", "!="]
        threshold:
          type: number
        isActive:
          type: boolean
        sensorId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /sensors/{sensorId}/rules:
    parameters:
      - name: sensorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Listar regras de um sensor
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de regras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
    post:
      summary: Criar regra para um sensor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rule'
      responses:
        '201':
          description: Regra criada
        '403':
          description: Acesso negado
  /rules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Atualizar regra
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rule'
      responses:
        '200':
          description: Regra atualizada
        '403':
          description: Acesso negado
    delete:
      summary: Excluir regra
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Regra excluída
        '403':
          description: Acesso negado
...